Satisfy desktop file size policy
Extract anchored-scroll contracts and target-centering geometry so routed navigation remains within the desktop source-size ratchet.
